UserService, ArtistService were added. DAO layer was modified.

+++ src/main/java/am/aca/wftartproject/dao/ItemDao.java
@@ -10,39 +10,39 @@ import java.util.List;
 public interface ItemDao {
 
     /**
-     * Add Item info for appropriate artist
+     * Adds Item info for appropriate artist
      *
      * @param artistID
      * @param item
      */
-    void addItem(int artistID, Item item);
+    void addItem(Long artistID, Item item);
 
     /**
-     * Update Item price by id
+     * Updates Item price by id
      *
      * @param id
      * @param price
      */
-    void updateItem(int id, double price);
+    void updateItem(Long id, double price);
 
     /**
-     * Delete Item by id
+     * Deletes Item by id
      *
      * @param id
      */
-    void deleteItem(int id);
+    void deleteItem(Long id);
 
     /**
-     * Find Item by id
+     * Finds Item by id
      *
      * @param id
      * @return
      */
-    Item findItem(int id);
+    Item findItem(Long id);
 
 
     /**
-     * Get recently added items
+     * Gets recently added items
      *
      * @param limit
      * @return
@@ -50,7 +50,7 @@ public interface ItemDao {
     List<Item> getRecentlyAddedItems(int limit);
 
     /**
-     * Get all items with the following title.
+     * Gets all items with the following title.
      *
      * @param title
      * @return
@@ -59,7 +59,7 @@ public interface ItemDao {
 
 
     /**
-     * Get all items with the following type.
+     * Gets all items with the following type.
      *
      * @param itemType
      * @return
